requirements.txt
ollama
requests 
beautifulsoup4 
mysql-connector-python 
click
selenium
ollama
flask
streamlit
flask-cors